server: switch to rocket (#41)
as mentioned in #42 we want to switch away from warp to a server
framework that is nicer to use. this pr reimplements the dev server in
rocket with feature parity, and much much less lines of code

there were a few things i had to do other than porting the logic 1:1

- move all `tokio::main` initializations into their specific
subcommands, because rocket needs its own special tokio runtime

ideally, we would simply turn main from `#[tokio::main]` into
`#[rocket::main]`, but because of our use of html5ever_rcdom, some dom
parsing code is unsafe to be sent between threads. see:
https://rust-lang.github.io/rust-clippy/master/index.html#future_not_send

- implement a custom error type that wraps `eyre::color_eyre` to allow
for more diverse error handling

by default in rocket, returning the Err side of a result will log it in
the server's console, but simply just return a status code 500 to the
user. this breaks our old error handling strat, so I had to
[implement](https://github.com/delan/autost/pull/41/commits/261806ac96ee075f8c981b49aa47d765a5daa692)
[an](https://github.com/delan/autost/pull/41/commits/02c57035dd962ddb1c169e936bfdb61fa04b4812)
[enum](https://github.com/delan/autost/pull/41/commits/36577514356d0bfd986c2d975f6bcd8ef3aeeab8)